bioAPI/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
biopython>=1.81
orjson>=3.9.0
pysam>=0.22.0
